software/backend/requirements.txt
# Flask Core Framework
Flask==2.3.3
Flask-RESTful==0.3.10

# Authentication & Security
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# WebSocket Support
Flask-SocketIO==5.3.6
python-socketio==5.9.0

# Database Drivers
pymongo==4.5.0
redis==5.0.1

# Background Task Processing
celery==5.3.4

# Data Processing & AI Integration
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2

# Environment & Configuration
python-dotenv==1.0.0

# Validation & Serialization
marshmallow==3.20.1
jsonschema==4.19.1

# HTTP Client (for external API calls)
requests==2.31.0

# Logging & Monitoring
structlog==23.1.0

# Development & Testing Dependencies
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
pytest-mock==3.11.1
black==23.7.0
flake8==6.0.0

# Production Server
gunicorn==21.2.0
eventlet==0.33.3

# Report Generation
reportlab==4.0.4
openpyxl==3.1.2

# Date/Time Utilities
python-dateutil==2.8.2

# Cryptography (for secure operations)
cryptography==41.0.4

# Memory profiling (optional)
memory-profiler==0.61.0